app/requirements.txt
falcon==1.4.1
gunicorn==19.7.1
PyYAML==3.12
dependency_injector==3.11.3
Pillow==5.2.0
piexif==1.0.13
PyMySQL==0.8.0
SQLAlchemy==1.2.2
requests==2.18.4
nose==1.3.7
alembic==0.9.9
click==6.0
boto3==1.7.4
botocore==1.10.4
python-dotenv==0.8.2
grpcio==1.15.0
PyJWT==1.6.4
